+++ dap-etl/extraction/src/test/scala/org/broadinstitute/monster/dap/CslbExtractionPipelineBuilderIntegrationSpec.scala
@@ -1,8 +1,8 @@
 package org.broadinstitute.monster.dap
 
 import java.time.{LocalDate, LocalTime, OffsetDateTime, ZoneOffset}
-
 import better.files.File
+import common.Args
 import org.broadinstitute.monster.common.PipelineBuilderSpec
 
 class CslbExtractionPipelineBuilderIntegrationSpec extends PipelineBuilderSpec[Args] {
