+++ dap-etl/extraction/src/test/scala/org/broadinstitute/monster/dap/HLESExtractionPipelineBuilderIntegrationSpec.scala
@@ -2,6 +2,7 @@ package org.broadinstitute.monster.dap
 
 import java.time.{LocalDate, LocalTime, OffsetDateTime, ZoneOffset}
 import better.files.File
+import common.Args
 import org.broadinstitute.monster.common.PipelineBuilderSpec
 import upack._
